Migrate test models to reusable test fixtures
Since we will be writing different routines for testing, it is better to
use the same test "models" in testing, so let's reuse the same code.
