+++ model_signing/serialization/fixtures_constants.py
@@ -1,0 +1,18 @@
+# Copyright 2024 The Sigstore Authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Constants used in test fixtures and tests. Not part of the public API."""
+
+KNOWN_MODEL_TEXT: bytes = b"This is a simple model"
+ANOTHER_MODEL_TEXT: bytes = b"This is another simple model"
